Boss switching sides mechanic
Made it so rock attacks despawn after reaching the bottom of the screen. I added a SwitchSides() method to the Hurricane which makes it change screen side. Also made the Hurricane loop attacks and only die when it runs out of health.
I made it so the health of the Hurricane is reduced each wave cycle, this is only for debugging and will not be in the final game.